+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
-## 0.1.1
+## 0.2.0
+
+- Code cleanup and added `--watch` 👀 command
+  ([#b6eaabd](https://github.com/micheledinelli/wifind/commit/b6eaabd7b73fc1515665727795c572a26d44a9e3), @micheledinelli)
+- Added version command and file `_version.py`
+  ([#36df0b2](https://github.com/micheledinelli/wifind/commit/36df0b20493c09e66dc2f3c83bf33295e298dfc0), @micheledinelli)
+
+## 0.1.0
 
 First version of the package